+++ servers/core/coursePhase/coursePhaseDTO/get_course_phase_prev_data.go
@@ -1,14 +1,14 @@
 package coursePhaseDTO
 
 import (
-	"github.com/niclasheun/prompt2.0/coursePhase/resolution"
+	"github.com/niclasheun/prompt2.0/coursePhase/resolution/resolutionDTO"
 	db "github.com/niclasheun/prompt2.0/db/sqlc"
 	"github.com/niclasheun/prompt2.0/meta"
 )
 
 type PrevCoursePhaseData struct {
-	PrevData    meta.MetaData           `json:"prevData"`
-	Resolutions []resolution.Resolution `json:"resolutions"`
+	PrevData    meta.MetaData              `json:"prevData"`
+	Resolutions []resolutionDTO.Resolution `json:"resolutions"`
 }
 
 func GetPrevCoursePhaseDataDTO(prevCoreData []byte, resolutions []db.GetPrevCoursePhaseDataResolutionRow) (PrevCoursePhaseData, error) {
@@ -17,9 +17,9 @@ func GetPrevCoursePhaseDataDTO(prevCoreData []byte, resolutions []db.GetPrevCour
 		return PrevCoursePhaseData{}, err
 	}
 
-	resolutionDTOs := make([]resolution.Resolution, 0, len(resolutions))
-	for _, resolutionDTO := range resolutions {
-		resolutionDTOs = append(resolutionDTOs, resolution.GetPhaseResolutionDTOFromDBModel(resolutionDTO))
+	resolutionDTOs := make([]resolutionDTO.Resolution, 0, len(resolutions))
+	for _, resolution := range resolutions {
+		resolutionDTOs = append(resolutionDTOs, resolutionDTO.GetPhaseResolutionDTOFromDBModel(resolution))
 	}
 
 	return PrevCoursePhaseData{
