+++ servers/core/coursePhase/coursePhaseParticipation/coursePhaseParticipationDTO/get_participations_with_resolution.go
@@ -1,8 +1,8 @@
 package coursePhaseParticipationDTO
 
-import "github.com/niclasheun/prompt2.0/coursePhase/resolution"
+import "github.com/niclasheun/prompt2.0/coursePhase/resolution/resolutionDTO"
 
 type CoursePhaseParticipationsWithResolutions struct {
 	Participations []GetAllCPPsForCoursePhase `json:"participations"`
-	Resolutions    []resolution.Resolution    `json:"resolutions"`
+	Resolutions    []resolutionDTO.Resolution `json:"resolutions"`
 }
